+++ argopy/xarray.py
@@ -691,6 +691,7 @@ class ArgoAccessor:
         """ Add TEOS10 variables to the dataset
 
         By default, adds: 'SA', 'CT', 'SIG0', 'N2', 'PV', 'PTEMP'
+        Other possible variables: 'CS'
         Relies on the gsw library.
 
         If one exists, the correct CF standard name will be added to the attrs.
@@ -716,6 +717,8 @@ class ArgoAccessor:
                 This variable has been regridded to the original pressure levels in the Dataset using a linear interpolation.
             * `"PTEMP"`
                 Adds a potential temperature variable
+            * `"CS"`
+                Adds a sound speed variable
             
         inplace: boolean, True by default
             If True, return the input :class:`xarray.Dataset` with new TEOS10 variables added as a new :class:`xarray.DataArray`
@@ -729,7 +732,7 @@ class ArgoAccessor:
             raise ModuleNotFoundError(
                 "This functionality requires the gsw library")
 
-        allowed = ['SA', 'CT', 'SIG0', 'N2', 'PV', 'PTEMP']
+        allowed = ['SA', 'CT', 'SIG0', 'N2', 'PV', 'PTEMP', 'CS']
         if any(var not in allowed for var in vlist):
             raise ValueError(f"vlist must be a subset of {allowed}, instead found {vlist}")
 
@@ -746,7 +749,6 @@ class ArgoAccessor:
         pres = this['PRES'].values
         lon = this['LONGITUDE'].values
         lat = this['LATITUDE'].values
-        f = lat
 
         # Coriolis
         f = gsw.f(lat)
@@ -781,6 +783,10 @@ class ArgoAccessor:
         # PV:
         if 'PV' in vlist:
             pv = f * n2 / gsw.grav(lat, pres)
+
+        # Sound Speed:
+        if 'CS' in vlist:
+            cs = gsw.sound_speed(sa, ct, pres)
 
         # Back to the dataset:
         that = []
@@ -823,6 +829,13 @@ class ArgoAccessor:
             PTEMP.attrs['standard_name'] = 'sea_water_potential_temperature'
             PTEMP.attrs['unit'] = 'degC'
             that.append(PTEMP)
+
+        if 'CS' in vlist:
+            CS = xr.DataArray(cs, coords=this['TEMP'].coords, name='CS')
+            CS.attrs['long_name'] = 'Speed of sound'
+            CS.attrs['standard_name'] = 'speed_of_sound_in_sea_water'
+            CS.attrs['unit'] = 'm/s'
+            that.append(CS)
 
         # Create a dataset with all new variables:
         that = xr.merge(that)