+++ docs/whats-new.rst
@@ -8,7 +8,7 @@ v0.1.8 (X XXX. 2021)
 
 **Features and front-end API**
 
-- Add Speed of sound computation to xarray accessor teos10 method.
+- Add Speed of sound computation to xarray accessor teos10 method. (:pr:`90`) by `G. Maze <http://www.github.com/gmaze>`_.
 
 v0.1.7 (4 Jan. 2021)
 -----------------------